feat: folder structure suggestion
