+++ src/components/Navbar/NavBar.style.scss
@@ -1,0 +1,20 @@
+.navbar {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  height: 100%;
+  padding: 0 2rem;
+
+  .logo {
+    text-decoration: none;
+    color: white;
+    font-weight: 500;
+  }
+}
+
+@media (min-width: 768px) {
+  .navbar {
+    padding: 0 3rem;
+  }
+}
