+++ src/components/Navbar/NavPages/NavPages.service.ts
@@ -1,0 +1,26 @@
+import { usePathname } from 'next/navigation';
+import { useEffect, useState } from 'react';
+
+export const useNavPagesService = () => {
+  const router = usePathname();
+
+  const [isMenuOpen, setIsMenuOpen] = useState(false);
+
+  const toggleMenu = () => {
+    setIsMenuOpen(!isMenuOpen);
+  };
+
+  useEffect(() => {
+    setIsMenuOpen(false);
+  }, [router]);
+
+  useEffect(() => {
+    setIsMenuOpen(false);
+  }, [router]);
+
+  return {
+    isMenuOpen,
+    toggleMenu,
+    setIsMenuOpen,
+  };
+};
