+++ src/components/Navbar/NavPages/NavPages.style.scss
@@ -1,0 +1,42 @@
+.desktop-menu {
+  display: none;
+}
+
+.menu-icon {
+  color: white;
+}
+
+.mobile-links {
+  position: absolute;
+  top: 100%;
+  left: 0;
+  right: 0;
+  background-color: rgba(0, 0, 0, 0.87);
+  border-radius: 5px;
+  padding: 1rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  transform: translateY(20px);
+  transition:
+    opacity 0.3s ease,
+    transform 0.3s ease;
+  opacity: 0;
+
+  &.active {
+    transform: translateY(0);
+    opacity: 1;
+  }
+}
+
+@media (min-width: 768px) {
+  .desktop-menu {
+    display: flex;
+    flex-direction: row;
+    column-gap: 1rem;
+  }
+
+  .mobile-menu {
+    display: none;
+  }
+}
