+++ src/components/Navbar/Pages/Pages.type.ts
@@ -1,0 +1,3 @@
+export interface IPages {
+  setIsMenuOpen: (val: boolean) => void;
+}
