--- src/styles/navbar.scss
@@ -1,79 +1,0 @@
-.navbar {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  height: 100%;
-  padding: 0 2rem;
-
-  .menu-icon {
-    color: white;
-  }
-
-  .logo {
-    text-decoration: none;
-    color: white;
-    font-weight: 500;
-  }
-
-  .navbar-link {
-    color: #fff;
-    text-decoration: none;
-    font-weight: 600;
-    padding: 0 0.5rem;
-    cursor: pointer;
-
-    &.active {
-      border-radius: 7px;
-      background-color: #75757538;
-    }
-  }
-
-  .desktop-menu {
-    display: none;
-  }
-
-  .mobile-links {
-    position: absolute;
-    top: 100%;
-    left: 0;
-    right: 0;
-    background-color: rgba(0, 0, 0, 0.87);
-    border-radius: 5px;
-    padding: 1rem;
-    display: flex;
-    flex-direction: column;
-    gap: 0.5rem;
-    transform: translateY(20px);
-    transition:
-      opacity 0.3s ease,
-      transform 0.3s ease;
-    opacity: 0;
-
-    &.active {
-      transform: translateY(0);
-      opacity: 1;
-    }
-
-    .navbar-link {
-      display: flex;
-      align-self: self-start;
-    }
-  }
-}
-
-@media (min-width: 768px) {
-  .navbar {
-    padding: 0 3rem;
-
-    .desktop-menu {
-      display: flex;
-      flex-direction: row;
-      column-gap: 1rem;
-    }
-
-    .mobile-menu {
-      display: none;
-    }
-  }
-}
